requirements.txt
aiogram>=3.17.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
apscheduler>=3.10.0
dateparser>=1.2.0
python-dateutil>=2.8.2
pytz>=2024.1
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
